Review #2: keep the gateway's served types (DB sweep item), safe_name event keys, positive-tolerance motors, headers aggregated, PVA dep dropped, mixin folded back
- P1: the option-list-is-enum rule change would have broken today's per-scan
  path for the 18 re-typed variables (the existing factories declare float;
  ophyd-async refuses float on an enum PV). The rule moves to GEECS-Core
  UNCHANGED; the 18 rows are recorded as a DB fix (variabletype='choice') so
  gateway and clients flip together. GeecsCAGateway back to a patch (0.20.3);
  PV_CONTRACT and its pin restored; GEECS-Core changelog says "no behaviour
  change".
- P2: a settable becomes a CaMotor only for a POSITIVE DB tolerance; 0/NULL
  is a plain CaSettable (exposure-like numerics never echo within a
  tolerance). The duplicated default-tolerance constant is gone.
- P2: ophyd names / event keys follow EVENT_SCHEMA.md (safe_name, lowercase);
  only the namespace binding keeps GEECS case. Children's _column_headers
  are aggregated onto the parent (run_wrapper reads top-level devices).
- P3: GeecsPvaGateway's dead geecs-ca-gateway dependency dropped (relocked);
  root CLAUDE.md graph updated.
- P3: the shot-monitor mixin extraction lost its second host when
  GeecsDevice was deleted — folded back into triggerable.py; only the
  datatypes= amendment remains there.
- P3 hygiene: unreachable int mapping removed, CLAUDE.md package layout,
  shared DocCollector test helper, version note.
