+++ GeecsCAGateway/tests/test_config_from_db.py
@@ -505,8 +505,9 @@ def test_from_geecs_experiment_all_variables(monkeypatch) -> None:
     assert {v.geecs_var for v in by_name["U_B"].variables} == {"Voltage"}
 
 
-def test_numeric_looking_option_list_is_an_enum() -> None:
-    """PV_CONTRACT §4: an option list is always enum, even with variabletype=numeric."""
+def test_variabletype_numeric_with_an_option_list_stays_float() -> None:
+    """PV_CONTRACT §4: variabletype wins over an option list (the served type
+    today; the 18 such Undulator rows are a DB fix — variabletype='choice')."""
     meta = [
         {
             "name": "FilterWheel",
@@ -520,5 +521,4 @@ def test_numeric_looking_option_list_is_an_enum() -> None:
     ]
     spec = DeviceSpec.from_db_metadata("D", "h", 1, meta)
     (var,) = spec.variables
-    assert var.dtype == "enum"
-    assert var.choices == ["1", "2", "3", "4", "5", "6"]
+    assert var.dtype == "float"
