+++ GeecsPvaGateway/CHANGELOG.md
@@ -8,7 +8,9 @@ Format follows [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 ### Changed
 
 - `effective_vartype` is imported from `geecs_core.db.variable_types`
-  (GEECS-Core 0.5.0) instead of `geecs_ca_gateway.config`. No behaviour change.
+  (GEECS-Core 0.5.0) instead of `geecs_ca_gateway.config`; the
+  `geecs-ca-gateway` path dependency — kept only for that helper — is
+  dropped. No behaviour change.
 
 ## [0.6.0] - 2026-09-05
