+++ GeecsPvaGateway/pyproject.toml
@@ -18,9 +18,6 @@ numpy = ">=2.0,<3.0"
 # camera servers (one instance per host, serving that host's cameras) instead
 # of centrally; the central CA gateway never proxies pixels (DESIGN.md).
 geecs-core = { path = "../GEECS-Core", develop = true }
-# Still needed beyond geecs-core for config helpers (effective_vartype —
-# the DB→dtype choice-descriptor rules shared with the CA gateway).
-geecs-ca-gateway = { path = "../GeecsCAGateway", develop = true }
 geecs-data-utils = { path = "../GEECS-Data-Utils", develop = true }
 
 [tool.poetry.group.dev.dependencies]
